get_avaible_serial_ports
