+++ src/ui/sidebar.rs
@@ -1,6 +1,5 @@
-use crate::{app, ui};
+use crate::{app, ui, util};
 use eframe::egui;
-use serialport::available_ports;
 
 pub fn select_dbc(
     app: &mut app::DAQApp,
@@ -86,24 +85,7 @@ pub fn show(app: &mut app::DAQApp, ctx: &egui::Context) {
                         }
                     });
                 if ui.button("🔄").clicked() {
-                    app.serial_ports = match available_ports() {
-                        Ok(ports) => ports
-                            .into_iter()
-                            .filter(|p| {
-                                let name = p.port_name.to_lowercase();
-                                if cfg!(target_os = "windows") {
-                                    name.starts_with("com")
-                                } else {
-                                    name.starts_with("/dev/tty.usbmodem")
-                                        || name.starts_with("/dev/ttyacm")
-                                }
-                            })
-                            .collect(),
-                        Err(err) => {
-                            log::error!("Failed to get ports: {err}");
-                            vec![]
-                        }
-                    };
+                    app.serial_ports = util::get_avaible_serial_ports();
                 }
             });
             if let Some(ref err) = app.connection_error {
